241223 프로젝트 수정 후 다시 반경 ### 변경사항 ### 1. 'io.spring.dependency-management' 버전 최적화 2. 인텔리제이 업그레이드 - 24년버전으로 3. 컨트롤러, dto를 application의 하위 프로젝트로 설정하고 프로젝트 구조를 최적화 4. logger는 java.util에서 slf4j으로 변경 5. .gitignore 수정 (불필요한 코드 제거) 6. 축약어로 적어놨던 패키지명을 풀네임으로 변경
