+++ api/src/main/java/com/tmarket/ApiApplication.java
@@ -1,0 +1,15 @@
+package com.tmarket;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+
+
+@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
+@ComponentScan(basePackages= {"com.tmarket.*", "com.tmarket.controller.*"})
+public class ApiApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(ApiApplication.class, args);
+    }
+}
